Version: 2014-12-24 ksh93v-
Sourced from:
a489176e4b3ae0c50f5b92f18d6f505cb8402b815d0e4a9b70f46887956f3a07  ast-open.2014-12-24.tgz
8852b9d37d5034e3780aeb5f963726381eeb4e08bb5bee1fbfa7e3f529c10e1b  INIT.2014-12-24.tgz

+++ src/cmd/ksh93/sh/xec.c
@@ -4143,7 +4143,7 @@ int sh_funscope_20120720(Shell_t *shp,int argn, char *argv[],int(*fun)(void*),vo
 	}
 	if(jmpval)
 		r=shp->exitval;
-	if(r>SH_EXITSIG && ((r&SH_EXITMASK)==SIGINT || ((r&SH_EXITMASK)==SIGQUIT)))
+	if(!sh_isstate(shp,SH_IOPROMPT) && r>SH_EXITSIG && ((r&SH_EXITMASK)==SIGINT || ((r&SH_EXITMASK)==SIGQUIT)))
 		kill(getpid(),r&SH_EXITMASK);
 	if(jmpval > SH_JMPFUN)
 	{
